requirements.txt
charset-normalizer==3.4.4
contourpy==1.3.3
cycler==0.12.1
Faker==40.1.2
fonttools==4.61.0
ipython-sql==0.5.0
kiwisolver==1.4.9
matplotlib==3.10.7
numpy==2.3.5
packaging==25.0
pillow==12.0.0
prettytable==3.10.2
psycopg2-binary==2.9.11
pyparsing==3.2.5
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
reportlab==4.4.9
six==1.17.0
sqlparse==0.5.0
tabulate==0.9.0
tzdata==2025.3
